Autonomous-Taxi-System/autonomous_taxi_system/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.21
PyMySQL==1.1.0
pandas==2.0.3
openpyxl==3.1.2
python-dotenv==1.0.0
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
email-validator==2.1.0.post1
reportlab==4.0.5
mysql-connector-python==8.0.26
flask-cors==5.0.1
Flask-SocketIO==5.5.1
